% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_installed.R
\name{is_installed}
\alias{is_installed}
\title{Check if a package is installed (in the terminal)}
\usage{
is_installed(pkg = NULL)
}
\arguments{
\item{pkg}{Character indicating the package to search.}
}
\value{
Logical of length one (i.e., TRUE or FALSE).
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
is_installed("npm")
is_installed("brew")

}
