% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encrypt_html.R
\name{protect_html}
\alias{protect_html}
\title{Encrypt HTML files}
\usage{
protect_html(html, password, output = NULL, title = NULL, instructions = NULL)
}
\arguments{
\item{html}{The address to an HTML file.}

\item{password}{The password used to protect the file.}

\item{output}{The address to an HTML file. If exists, the function will overwrite it.}

\item{title}{Title for the password box.}

\item{instructions}{Instructions for end user. Could be guidelines related to the password it self.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
Add a layer of protection to your HTML files. \code{protect_html()}
adds a password request layer before entering the final html file.
}
\section{About the algorithm}{


This uses under the hood the package staticrypt (\href{https://github.com/robinmoisson/staticrypt}{see GitHub repo}),
from \href{https://github.com/robinmoisson}{Robin Moisson}. For extra documentation about the JavaScript code see the
\href{https://github.com/robinmoisson}{repository} from his GitHub.
}

