% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlockdown-package.R
\docType{package}
\name{rlockdown-package}
\alias{rlockdown}
\alias{rlockdown-package}
\title{Password layer for HTML files}
\description{
Protect standalone static html files using no server-side comunication.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/matcasti/rlockdown}
  \item Report bugs at \url{https://github.com/matcasti/rlockdown/issues}
}

}
\author{
\strong{Maintainer}: Matías Castillo Aguilar \email{matcasti@umag.cl} (\href{https://orcid.org/0000-0001-7291-247X}{ORCID})

}
\keyword{internal}
